stonk-request/requirements.txt
google-cloud-pubsub>=2.6.^
